+++ gamemodes/terrortown/entities/weapons/weapon_ttt_beacon.lua
@@ -12,7 +12,7 @@ if CLIENT then
 	SWEP.PrintName = "beacon_name"
 	SWEP.Slot = 6
 
-	SWEP.ShowViewModel = false
+	SWEP.ShowDefaultViewModel = false
 
 	SWEP.EquipMenuData = {
 		type = "item_weapon",