+++ gamemodes/terrortown/entities/weapons/weapon_ttt_binoculars.lua
@@ -14,7 +14,7 @@ if CLIENT then
 	SWEP.PrintName = "binoc_name"
 	SWEP.Slot = 7
 
-	SWEP.ShowViewModel = false
+	SWEP.ShowDefaultViewModel = false
 
 	SWEP.EquipMenuData = {
 		type = "item_weapon",