+++ gamemodes/terrortown/entities/weapons/weapon_ttt_decoy.lua
@@ -8,7 +8,7 @@ else -- CLIENT
 	SWEP.PrintName = "decoy_name"
 	SWEP.Slot = 7
 
-	SWEP.ShowViewModel = false
+	SWEP.ShowDefaultViewModel = false
 
 	SWEP.EquipMenuData = {
 		type = "item_weapon",