+++ gamemodes/terrortown/entities/weapons/weapon_ttt_defuser.lua
@@ -13,7 +13,7 @@ if CLIENT then
 	SWEP.PrintName = "defuser_name"
 	SWEP.Slot = 7
 
-	SWEP.ShowViewModel = false
+	SWEP.ShowDefaultViewModel = false
 
 	SWEP.EquipMenuData = {
 		type = "item_weapon",